requirements.txt
Pillow~=8.4.0
Flask~=2.0.2
tensorflow>=2.0.0
numpy~=1.21.2
pathlib~=1.0.1